Humidity Sensor/requirements.txt
influxdb
paho_mqtt
pyserial
requests
